chore: remove facade/acl_commands.h
