+++ src/server/acl/user_registry.cc
@@ -8,7 +8,6 @@
 #include <mutex>
 
 #include "base/flags.h"
-#include "facade/acl_commands_def.h"
 #include "facade/facade_types.h"
 #include "server/acl/acl_commands_def.h"
