+++ src/server/conn_context.cc
@@ -8,7 +8,6 @@
 
 #include "base/logging.h"
 #include "common/heap_size.h"
-#include "facade/acl_commands_def.h"
 #include "facade/reply_builder.h"
 #include "server/acl/acl_commands_def.h"
 #include "server/channel_store.h"
